Update setsockopt.c to accept strings containing null
In the master version of matlab-zmq, a ZMQ_SUBSCRIBE message filter is treated as a SOPT_STRING, which is truncated at the first null terminating character on Line 56. This is not the correct behavior, because message filters are allowed to contain binary data including null characters. Since this is a MEX interface, we can use msGetNumberOfElements to get the correct optLen.